Renamed native files for clarity
